feat(extension) Introduce the `wasm_lazy_byte_array structure.
When calling `wasm_read_bytes($file_path)`, instead of reading the
file and allocating a `wasmer_byte_array`, a `wasm_lazy_byte_array` is
called. The file is not read at this moment.

This `wasm_lazy_byte_array` reads the file when its `get_bytes`
function is called. It is called in a single place: In
`wasm_bytes_from_resource`, i.e. when another function needs the
bytes. So the file and its bytes are read at the last moment, lazily.

Why? Because we can register a persistent module resource, but the
bytes were always read. With this strategy, the bytes are read only
when required by a function, and not when `wasm_read_bytes` are
effectively run.